refactor: externalize all functions
put create_csv() and extract_csv_info() in csv_management/csv_management.py
put filter_by_category() and filter_by_type() in filter/filter.py
