+++ scrap_jobs.py
@@ -1,15 +1,21 @@
 import pathlib
 import os
-import csv
 
 from scrapy.crawler import CrawlerProcess
 from spiders.django_jobs_spider import DjangojobsSpider
 from spiders.python_jobs_spider import PythonjobsSpider
 
-from csv_management.csv_creator import create_csv
+from csv_management.csv_management import extract_csv_info
+from filter.filter import filter_by_category, filter_by_type
 
 
 def main():
+    """Main function :
+    - Crawl spiders simultaneously.
+    - Move the created csv files in a data folder.
+    - Filter the jobs by category and type,
+     creating a file by type and category, in a dedicated folder.
+    """
 
     # Crawling Spiders
     process = CrawlerProcess()
@@ -17,44 +23,15 @@ def main():
     process.crawl(PythonjobsSpider)
     process.start()
 
-    # File organizing
+    # Organize Files
     os.makedirs(pathlib.Path.cwd() / 'data', exist_ok=True)
     os.replace(pathlib.Path.cwd() / 'djangojobs.csv', pathlib.Path.cwd() / 'data' / 'djangojobs.csv')
     os.replace(pathlib.Path.cwd() / 'pythonjobs.csv', pathlib.Path.cwd() / 'data' / 'pythonjobs.csv')
 
-    # extract info from csv
-    lists_of_all_information = []
-    with open(pathlib.Path.cwd() / 'data' / 'pythonjobs.csv', 'r', encoding='utf-8') as f:
-        content = csv.reader(f)
-        for row in content:
-            lists_of_all_information.append(row)
-    csv_headers = lists_of_all_information.pop(0)
-
-    # filter by category
-    list_of_categories = ["Data_Analyst", "Developer_Engineer", "Manager_Executive", "Other", "Researcher_Scientist"]
-    for category in list_of_categories:
-        lists_of_jobs_in_category = []
-        for job in lists_of_all_information:
-            if category[:4] in job[8]:
-                lists_of_jobs_in_category.append(job)
-        if lists_of_jobs_in_category:
-            create_csv("sorted_by_category", f"Category_{category}.csv", csv_headers, lists_of_jobs_in_category)
-
-    # filter by type
-    list_of_types = [
-        'Back end', 'Big Data', 'Cloud', 'Database', 'Evangelism', 'Finance',
-        'Front end', 'Fundraising', 'Image Processing', 'Integration', 'Lead',
-        'Machine Learning', 'Management', 'Numeric processing', 'Operations',
-        'Systems', 'Testing', 'Text Processing', 'Web'
-    ]
-    for job_type in list_of_types:
-        lists_of_jobs_in_types = []
-        for job in lists_of_all_information:
-            if job_type in job[7]:
-                lists_of_jobs_in_types.append(job)
-        if lists_of_jobs_in_types:
-            filename_job_type = job_type.replace(" ", "_")
-            create_csv("sorted_by_type", f"Type_{filename_job_type}.csv", csv_headers, lists_of_jobs_in_types)
+    # Filter
+    lists_of_all_information, csv_headers = extract_csv_info()
+    filter_by_category(csv_headers, lists_of_all_information)
+    filter_by_type(csv_headers, lists_of_all_information)
 
 
 if __name__ == '__main__':
